refactor: change filenames from cpuplot to sysinfo
